# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Yapily Integration API
  version: 1.0.0
  description: API documentation for Yapily Integration service
servers:
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SyncStatus:
      type: object
      properties:
        accountId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
        total:
          type: number
        error:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time

# Remove or modify the global security requirement
security: [] # Remove this or comment it out
paths:
...
  /api/monitor/sync/{syncId}:
    get:
      tags: [Monitor]
      summary: Get sync status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: syncId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Sync status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatus'
...